Completed AutoMgr refactor, removed unused files
